+++ docs/docs/plugin/fixtures.md
@@ -1,0 +1,8 @@
+# Pytest fixtures
+
+::: pdm.pytest
+    options:
+      show_source: false
+      show_root_heading: false
+      show_root_toc_entry: false
+      heading_level: 2
