+++ docs/docs/plugin/reference.md
@@ -1,7 +1,13 @@
 # API Reference
 
 ::: pdm.core.Core
-    rendering:
+    options:
+      show_root_heading: yes
+      show_source: false
+      heading_level: 2
+
+::: pdm.core.Project
+    options:
       show_root_heading: yes
       show_source: false
       heading_level: 2
@@ -11,5 +17,5 @@
 _New in version 1.12.0_
 
 ::: pdm.signals
-    rendering:
+    options:
       heading_level: 3
