+++ news/1594.feature.md
@@ -1,0 +1,1 @@
+- Expose PDM fixtures as a `pytest` plugin `pdm.pytest` for plugin developers.
